+++ src/schemaGenerator.ts
@@ -7,6 +7,7 @@
 // a bunch of utility functions into a separate utitlities folder, one file per function.
 
 import {
+  GraphQLEnumType,
   DocumentNode,
   parse,
   print,
@@ -369,6 +370,19 @@ function addResolveFunctionsToSchema(
 
       if (type instanceof GraphQLScalarType) {
         type[fieldName] = resolveFunctions[typeName][fieldName];
+        return;
+      }
+
+      if (type instanceof GraphQLEnumType) {
+        // TODO: Remove once https://github.com/DefinitelyTyped/DefinitelyTyped/pull/21786
+        // is inside NPM
+        if (!(type as any).isValidValue(fieldName)) {
+          throw new SchemaError(
+              `${typeName}.${fieldName} was defined in resolvers, but enum is not in schema`,
+          );
+        }
+
+        type.getValue(fieldName)['value'] = resolveFunctions[typeName][fieldName];
         return;
       }